Refactor: Product, Wallet Context 파일 분리
Related to: #15
